+++ CHANGELOG.md
@@ -8,12 +8,17 @@
 * Docs on kyc recovery
 * `/v3/accounts/{id}/sales/{sale_id}` endpoint to get sale by id, if account is allowed to participate
 * `GetSaleForAccount` to handle `/v3/accounts/{id}/sales/{sale_id}`
+* `Asset` to withdraw request relationships
+* Filter by asset for `/v3/create_withdraw_requests` and `/v3/create_withdraw_requests/{id}`
+* Ability to include asset in responses`/v3/create_withdraw_requests` and `/v3/create_withdraw_requests/{id}`
 
 ### Fixed
 
 * docs key value u32 value
 * add `trailing_digits_count` to asset creation request
 * Panic on `/v3/history` handling `ManageAccountSpecificRuleOp` details
+* Own sales appearing on `/v3/account/{id}/sales`
+* Legacy withdraw requests endpoint asset filter
 
 ### Changed